Create SystemAbstractions project
- Enable other projects to use the system abstractions formerly limited to Axe.Windows.Automation
	- Allow System calls to be mocked so Fakes can be replaced
- Use .Net Standard to ensure broadest future compatibility
