+++ src/Automation/Automation.csproj
@@ -58,21 +58,13 @@
       <Project>{7FE58829-560E-4142-BF81-291A66D3844B}</Project>
       <Name>Rules</Name>
     </ProjectReference>
+    <ProjectReference Include="..\SystemAbstractions\SystemAbstractions.csproj">
+      <Project>{7e616c64-ead7-458c-92ac-b50643e64d38}</Project>
+      <Name>SystemAbstractions</Name>
+    </ProjectReference>
   </ItemGroup>
   <ItemGroup>
     <Compile Include="$(TEMP)\AxeWindowsVersionInfo.cs" />
-    <Compile Include="Abstractions\ISystem.cs" />
-    <Compile Include="Abstractions\ISystemDateTime.cs" />
-    <Compile Include="Abstractions\ISystemFactory.cs" />
-    <Compile Include="Abstractions\ISystemIO.cs" />
-    <Compile Include="Abstractions\ISystemIODirectory.cs" />
-    <Compile Include="Abstractions\ISystemEnvironment.cs" />
-    <Compile Include="Concretions\System.cs" />
-    <Compile Include="Concretions\SystemDateTime.cs" />
-    <Compile Include="Concretions\SystemEnvironment.cs" />
-    <Compile Include="Concretions\SystemFactory.cs" />
-    <Compile Include="Concretions\SystemIO.cs" />
-    <Compile Include="Concretions\SystemIODirectory.cs" />
     <Compile Include="Data\Config.cs" />
     <Compile Include="Data\ElementInfo.cs" />
     <Compile Include="Data\OutputFile.cs" />
