+++ src/AxeWindows.sln
@@ -1,7 +1,7 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.28010.2050
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.29102.190
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Actions", "Actions\Actions.csproj", "{B2FA5150-470A-4864-A191-7BCAE39DF35A}"
 EndProject
@@ -58,6 +58,10 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RulesTests", "RulesTest\Rul
 	EndProjectSection
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CI", "CI\CI.csproj", "{32273060-8BBE-4BB6-9198-50F9C680BE60}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SystemAbstractions", "SystemAbstractions\SystemAbstractions.csproj", "{7E616C64-EAD7-458C-92AC-B50643E64D38}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SystemAbstractionsTests", "SystemAbstractionsTests\SystemAbstractionsTests.csproj", "{683CBD73-C7B8-4662-A9DE-96E434E167D5}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -297,6 +301,30 @@ Global
 		{32273060-8BBE-4BB6-9198-50F9C680BE60}.Release|x64.Build.0 = Release|Any CPU
 		{32273060-8BBE-4BB6-9198-50F9C680BE60}.Release|x86.ActiveCfg = Release|Any CPU
 		{32273060-8BBE-4BB6-9198-50F9C680BE60}.Release|x86.Build.0 = Release|Any CPU
+		{7E616C64-EAD7-458C-92AC-B50643E64D38}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{7E616C64-EAD7-458C-92AC-B50643E64D38}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{7E616C64-EAD7-458C-92AC-B50643E64D38}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{7E616C64-EAD7-458C-92AC-B50643E64D38}.Debug|x64.Build.0 = Debug|Any CPU
+		{7E616C64-EAD7-458C-92AC-B50643E64D38}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{7E616C64-EAD7-458C-92AC-B50643E64D38}.Debug|x86.Build.0 = Debug|Any CPU
+		{7E616C64-EAD7-458C-92AC-B50643E64D38}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{7E616C64-EAD7-458C-92AC-B50643E64D38}.Release|Any CPU.Build.0 = Release|Any CPU
+		{7E616C64-EAD7-458C-92AC-B50643E64D38}.Release|x64.ActiveCfg = Release|Any CPU
+		{7E616C64-EAD7-458C-92AC-B50643E64D38}.Release|x64.Build.0 = Release|Any CPU
+		{7E616C64-EAD7-458C-92AC-B50643E64D38}.Release|x86.ActiveCfg = Release|Any CPU
+		{7E616C64-EAD7-458C-92AC-B50643E64D38}.Release|x86.Build.0 = Release|Any CPU
+		{683CBD73-C7B8-4662-A9DE-96E434E167D5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{683CBD73-C7B8-4662-A9DE-96E434E167D5}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{683CBD73-C7B8-4662-A9DE-96E434E167D5}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{683CBD73-C7B8-4662-A9DE-96E434E167D5}.Debug|x64.Build.0 = Debug|Any CPU
+		{683CBD73-C7B8-4662-A9DE-96E434E167D5}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{683CBD73-C7B8-4662-A9DE-96E434E167D5}.Debug|x86.Build.0 = Debug|Any CPU
+		{683CBD73-C7B8-4662-A9DE-96E434E167D5}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{683CBD73-C7B8-4662-A9DE-96E434E167D5}.Release|Any CPU.Build.0 = Release|Any CPU
+		{683CBD73-C7B8-4662-A9DE-96E434E167D5}.Release|x64.ActiveCfg = Release|Any CPU
+		{683CBD73-C7B8-4662-A9DE-96E434E167D5}.Release|x64.Build.0 = Release|Any CPU
+		{683CBD73-C7B8-4662-A9DE-96E434E167D5}.Release|x86.ActiveCfg = Release|Any CPU
+		{683CBD73-C7B8-4662-A9DE-96E434E167D5}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
