+++ src/SystemAbstractions/Abstractions/ISystem.cs
@@ -1,13 +1,13 @@
 ﻿// Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 using System;
-using System.IO;
 
-namespace Axe.Windows.Abstractions
+namespace SystemAbstractions
 {
-    internal interface ISystemIODirectory
+    public interface ISystem
     {
-        DirectoryInfo CreateDirectory(string path);
-        bool Exists(string path);
+        ISystemDateTime DateTime { get; }
+        ISystemEnvironment Environment { get; }
+        ISystemIO IO { get; }
     } // interface
 } // namespace
