+++ src/SystemAbstractions/Abstractions/ISystemDateTime.cs
@@ -1,12 +1,11 @@
 ﻿// Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
-using Axe.Windows.Abstractions;
 using System;
 
-namespace Axe.Windows.Concretions
+namespace SystemAbstractions
 {
-    internal class SystemDateTime : ISystemDateTime
+    public interface ISystemDateTime
     {
-        public DateTime Now => DateTime.Now;
-    } // class
+        DateTime Now { get; }
+    } // interface
 } // namespace
