+++ src/SystemAbstractions/Abstractions/ISystemFactory.cs
@@ -1,8 +1,7 @@
 ﻿// Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
-using Axe.Windows.Abstractions;
 
-namespace Axe.Windows.Abstractions
+namespace SystemAbstractions
 {
     internal interface ISystemFactory
     {
