+++ src/SystemAbstractions/Concretions/SystemIO.cs
@@ -1,9 +1,8 @@
 ﻿// Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
-using Axe.Windows.Abstractions;
 using System;
 
-namespace Axe.Windows.Concretions
+namespace SystemAbstractions
 {
     class SystemIO : ISystemIO
     {
