+++ src/SystemAbstractions/Concretions/SystemIODirectory.cs
@@ -1,10 +1,9 @@
 ﻿// Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
-using Axe.Windows.Abstractions;
 using System;
 using System.IO;
 
-namespace Axe.Windows.Concretions
+namespace SystemAbstractions
 {
     internal class SystemIODirectory : ISystemIODirectory
     {
