+++ src/SystemAbstractions/SystemAbstractions.csproj
@@ -1,0 +1,8 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>netstandard2.0</TargetFramework>
+    <AssemblyName>Axe.Windows.SystemAbstractions</AssemblyName>
+  </PropertyGroup>
+
+</Project>
